Made (default) maximum hits to retrieve/count configurable.
See the BlackLab Server overview on the GitHub wiki for details, but
defaults and maximum values can be configured in blacklab-server.json
and uses can specify values different from the defaults in the URL
(maxretrieve and maxcount), as long as they don't exceed the configured
maximum values. stoppedCountingHits / -RetrievingHits in output
indicates whether or not the configured maximum values were reached.