Added code for logged-in users creating their own new indices.
Note that the login system doesn't fully work yet, so this experimental
functionality is not actually available. When it's done, it will be
turned off by default for security reasons. New indices will always be
created inside a private directory for that user. Input is strictly
validated to avoid security vulnerabilities.