Ignore incorrect sort values instead of issuing an error.
The problem is that applications might remember your preferred sort and
pass it with subsequent searches, even if that particular sort cannot be
performed on that type of search.

In effect we've decided that it's worse to force every client to
explicitly validate every sort value / search type combination than to
ignore the occasional sort requests that doesn't match the search type.