Worked on indexing operations (experimental).
Logged-in users can create and add to their private indices. The "create
index" operation was changed from PUT /bls/indexName to POST /bls
because we need to pass more parameters (optional display name and
document format). Adding data to an index now closes the existing
Searcher and removes searches for this index from the cache.