Changes to how userId is sanitized for URL and directory name.
In the (experimental, off by default) private corpora system, user can
create, add data to and delete corpora. This works with their userId,
which is an arbitrary string returned by the authentication class in
use. This string needs to be sanitized to make a safe directory name and
to be usable in URLs. The first is done by throwing out anything except
letters and numbers and appending the MD5 hash of the original userId.
The second (URL sanitization) allows more special characters, e.g. email
adresses are acceptable as-is.