Switched to a different approach for load management.
Instead of applying rules to jobs directly, we instead sort the jobs on
"worthiness" and allow the top X searches to run, pause the next Y
searches, and abort any searches that are left. X and Y are configurable
in blacklab-server.json.

The worthiness heuristic is a bit complex, but takes into account how
long a job is running (jobs are given a while to execute, then their
worthiness is lowered as they reach a certain age) and whether it is an
actual search or just a "total results counter", which is deemed less
important. See Job.compareTo() for the complete logic.